refactored the 4 validation functions into 1
